coredns: reply with NX if there are no servers
There might be no DNS servers configured because `/etc/resolv.conf` is
empty or doesn't exist. Not replying causes the client to try the next
server instead of trying to aardvark with the next search domain.

In practice, this caused a client that has a search-domain like
`.home.arpa` in it's resolv.conf to not retry with `.` because from it's
perspective the server is unreachable.
Thus, the domains in `/run/containers/networks/aardvark-dns/podman` were
not resolved successfully.